Support the loops that doesn't exit at it's loop head.
A loop may exit at some branch instructions that is not the entry
point of the loop.  A loop can even exist at more than one location.
For example, a while-loop can have mutliple 'break' instructions in
it.

The new method handles loops by finding all instructions between two
executions of the same instruction and finding branch instructions, in
a loop, having one its branch staying in the loop and the other
leaving the loop.

Signed-off-by: Thinker Lee <thinker.li@gmail.com>